refactor: batch のジョブロジックを internal/app/batch へ移設
- cmd/batch/main.go を batch.Run を呼ぶだけの薄いラッパーに変更
- ジョブのオーケストレーション（candles/logo の DI・実行）を internal/app/batch パッケージへ移動
- ディスパッチャを公開関数 Run にリネームし、ヘルパー（failureRater/shouldFailExit/parse*）も同梱
- main_test.go を batch_test.go として通常パッケージのテストへ移設
