+++ internal/app/batch/logo.go
@@ -1,0 +1,63 @@
+package batch
+
+import (
+	"context"
+	"log/slog"
+	"time"
+
+	"stock_backend/internal/app/di"
+	symbollistadapters "stock_backend/internal/feature/symbollist/adapters"
+	symbollistusecase "stock_backend/internal/feature/symbollist/usecase"
+	"stock_backend/internal/platform/db"
+	"stock_backend/internal/shared/clientratelimit"
+)
+
+// runLogoIngest は TwelveData からロゴURLを取り込み、終了コード（0 or 1）を返す。
+func runLogoIngest() int {
+	sqlDB, err := db.OpenSQL()
+	if err != nil {
+		slog.Error("DB open failed", "error", err)
+		return 1
+	}
+	defer func() {
+		if err := sqlDB.Close(); err != nil {
+			slog.Warn("failed to close sqlDB", "error", err)
+		}
+	}()
+	logoProvider := di.NewMarket()
+	symbolRepo := symbollistadapters.NewSymbolRepository(sqlDB)
+	rateLimiter := clientratelimit.NewRateLimiter(rateLimitPerMinute, time.Minute)
+	uc := symbollistusecase.NewLogoIngestUsecase(logoProvider, symbolRepo, rateLimiter)
+
+	timeoutHours := parseTimeoutHours("LOGO_INGEST_TIMEOUT_HOURS", 3)
+	ctx, cancel := context.WithTimeout(context.Background(), time.Duration(timeoutHours)*time.Hour)
+	defer cancel()
+
+	maxFailureRate := parseMaxFailureRate("LOGO_INGEST_MAX_FAILURE_RATE", defaultMaxFailureRate)
+
+	start := time.Now()
+	result, err := uc.IngestAll(ctx)
+	duration := time.Since(start)
+
+	slog.Info("logo ingest summary",
+		"total", result.Total,
+		"succeeded", result.Succeeded,
+		"failed", result.Failed,
+		"failure_rate", result.FailureRate(),
+		"duration", duration.String(),
+	)
+
+	if err != nil {
+		slog.Error("logo ingest aborted by fatal error", "error", err)
+		return 1
+	}
+	if shouldFailExit(result, maxFailureRate) {
+		slog.Error("logo ingest failure rate exceeded threshold",
+			"failure_rate", result.FailureRate(),
+			"threshold", maxFailureRate,
+		)
+		return 1
+	}
+	slog.Info("logo ingest ok")
+	return 0
+}
